Update code from GitLab

+++ migrations/1700050939537422_stocks.postgres.up.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS stocks (
     exception BOOLEAN default false,
     amount INTEGER NOT NULL,
     vat_percentage INTEGER,
-    net_price INTEGER,
+    net_price FLOAT,
     organization_unit_id INTEGER NOT NULL,
     created_at TIMESTAMP,
     updated_at TIMESTAMP
